move BrowserType to browserType decode
