+++ static/app/views/insights/browser/webVitals/queries/useTransactionWebVitalsQuery.tsx
@@ -1,7 +1,7 @@
 import type {Sort} from 'sentry/utils/discover/fields';
-import type {BrowserType} from 'sentry/views/insights/browser/webVitals/components/browserTypeSelector';
 import {useTransactionWebVitalsScoresQuery} from 'sentry/views/insights/browser/webVitals/queries/storedScoreQueries/useTransactionWebVitalsScoresQuery';
 import type {WebVitals} from 'sentry/views/insights/browser/webVitals/types';
+import type {BrowserType} from 'sentry/views/insights/browser/webVitals/utils/queryParameterDecoders/browserType';
 
 type Props = {
   browserType?: BrowserType;
